+++ js/forms-controller.js
@@ -1,6 +1,3 @@
-const adForm = document.querySelector('.ad-form');
-const addressField = adForm.querySelector('#address');
-
 const disableForm = (form, className) => {
   form.classList.add(`${className}--disabled`);
   for (let i = 0; i < form.children.length; i++) {
@@ -15,8 +12,4 @@ const enableForm = (form, className) => {
   }
 };
 
-const setAddresValue = ({lat, lng}) => {
-  addressField.value = `${lat.toFixed(5)}, ${lng.toFixed(5)}`;
-};
-
-export { enableForm, disableForm, setAddresValue};
+export { enableForm, disableForm };
